+++ examples/data-sources/iosxe_community_list_expanded/data-source.tf
@@ -1,0 +1,3 @@
+data "iosxe_community_list_expanded" "example" {
+  name = "CLE1"
+}