+++ examples/data-sources/iosxe_community_list_standard/data-source.tf
@@ -1,0 +1,3 @@
+data "iosxe_community_list_standard" "example" {
+  name = "CLS1"
+}