+++ examples/resources/iosxe_community_list_expanded/import.sh
@@ -1,0 +1,1 @@
+terraform import iosxe_community_list_expanded.example "Cisco-IOS-XE-native:native/ip/Cisco-IOS-XE-bgp:community-list/expanded=CLE1"