+++ examples/resources/iosxe_community_list_standard/resource.tf
@@ -1,0 +1,5 @@
+resource "iosxe_community_list_standard" "example" {
+  name           = "CLS1"
+  deny_entries   = ["65000:500"]
+  permit_entries = ["65000:501"]
+}